requirements.txt
opencv-python
keras
tensorflow
numpy
pillow
